feat: 노선에 등록된 역 삭제 방지 기능 추가

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -3,6 +3,7 @@ package wooteco.subway.station.application;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
+import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
@@ -10,10 +11,13 @@ import wooteco.subway.station.dto.StationResponse;
 
 @Service
 public class StationService {
-    private StationDao stationDao;
 
-    public StationService(StationDao stationDao) {
+    private final StationDao stationDao;
+    private final SectionDao sectionDao;
+
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public StationResponse saveStation(StationRequest stationRequest) {
@@ -38,6 +42,9 @@ public class StationService {
     }
 
     public void deleteStationById(Long id) {
+        if (sectionDao.existsByStationId(id)) {
+            throw new IllegalArgumentException("이미 노선에 등록되어있는 역은 지울 수 없습니다.");
+        }
         stationDao.deleteById(id);
     }
 }
